Some improvements on the z-index of markers
But not easy to handle as Leaflet Circle objects are in SVG and
therefore cannot have a pre-defined z-index. So, z-order is based on
the order in which the elements were added to the map and as some of
them take more time to load than other, we can sometimes cannot decide
of the order.